feat(file_system): Add APIs for getting file paths and refactor implementation * Update `espp::FileSystem` to support new APIs for setting read_only and grow_on_mount before mounting * Update `espp::FileSystem` to move string serialization of permissions and file times to separate reusable functions * Add `espp::FileSystem::get_files_in_path(path, include_directories, recursive)` method to allow callers to operate on files / directories more effectively than having to do string parsing * Refactored to move bulk of implementation of `FileSystem` into cpp
