+++ components/file_system/CMakeLists.txt
@@ -1,3 +1,4 @@
 idf_component_register(
   INCLUDE_DIRS "include"
+  SRC_DIRS "src"
   REQUIRES base_component esp_littlefs spi_flash)
