+++ components/file_system/example/partitions.csv
@@ -2,4 +2,4 @@
 nvs,      data, nvs,     0x9000,  0x6000
 phy_init, data, phy,     0xf000,  0x1000
 factory,  app,  factory, 0x10000, 2M
-littlefs, data, spiffs,         , 2M
+littlefs, data, spiffs,         , 1M
