+++ components/file_system/include/file_system.hpp
@@ -45,6 +45,26 @@ namespace espp {
 /// \snippet file_system_example.cpp file_system std filesystem example
 class FileSystem : public BaseComponent {
 public:
+  /// @brief Set whether to mount the file system as read only
+  /// @param read_only Whether the file system is mounted as read only
+  /// @note This only has an effect if called before the file system is mounted,
+  ///       i.e. before the first call to get()
+  static void set_mount_as_read_only(bool read_only) { read_only_ = read_only; }
+
+  /// @brief Get whether the file system is mounted as read only
+  /// @return Whether the file system is mounted as read only
+  static bool is_mount_as_read_only() { return read_only_; }
+
+  /// @brief Set whether to grow the file system on mount
+  /// @param grow_on_mount Whether to grow the file system on mount
+  /// @note This only has an effect if called before the file system is mounted,
+  ///       i.e. before the first call to get()
+  static void set_grow_on_mount(bool grow_on_mount) { grow_on_mount_ = grow_on_mount; }
+
+  /// @brief Get whether the file system was grown on mount
+  /// @return Whether the file system was grown on mount
+  static bool is_grow_on_mount() { return grow_on_mount_; }
+
   /// @brief Access the singleton instance of the file system
   /// @return Reference to the file system instance
   static FileSystem &get() {
@@ -74,27 +94,15 @@ public:
 
   /// @brief Get the amount of free space on the file system
   /// @return The amount of free space in bytes
-  size_t get_free_space() {
-    size_t total, used;
-    esp_littlefs_info(get_partition_label(), &total, &used);
-    return total - used;
-  }
+  size_t get_free_space() const;
 
   /// @brief Get the total amount of space on the file system
   /// @return The total amount of space in bytes
-  size_t get_total_space() {
-    size_t total, used;
-    esp_littlefs_info(get_partition_label(), &total, &used);
-    return total;
-  }
+  size_t get_total_space() const;
 
   /// @brief Get the amount of used space on the file system
   /// @return The amount of used space in bytes
-  size_t get_used_space() {
-    size_t total, used;
-    esp_littlefs_info(get_partition_label(), &total, &used);
-    return used;
-  }
+  size_t get_used_space() const;
 
   /// @brief Get a human readable string for a byte size
   /// @details
@@ -103,18 +111,11 @@ public:
   /// https://en.cppreference.com/w/cpp/filesystem/file_size
   /// @param bytes The byte size
   /// @return The human readable string
-  std::string human_readable(size_t bytes) {
-    int i{};
-    float mantissa = bytes;
-    for (; mantissa >= 1024.; mantissa /= 1024., ++i) {
-    }
-    mantissa = std::ceil(mantissa * 10.) / 10.;
-    return i == 0 ? fmt::format("{}", bytes) : fmt::format("{:.3f}{}", mantissa, "BKMGTPE"[i]);
-  }
+  std::string human_readable(size_t bytes) const;
 
   /// @brief Get the partition label
   /// @return The partition label
-  const char *get_partition_label() { return CONFIG_ESPP_FILE_SYSTEM_PARTITION_LABEL; }
+  const char *get_partition_label() const { return CONFIG_ESPP_FILE_SYSTEM_PARTITION_LABEL; }
 
   /// @brief Get the mount point
   /// @details
@@ -122,14 +123,43 @@ public:
   /// It is the root directory of the partition with the partition label.
   /// @see get_root_path() and get_partition_label()
   /// @return The mount point
-  std::string get_mount_point() { return "/" + std::string{get_partition_label()}; }
+  std::string get_mount_point() const { return "/" + std::string{get_partition_label()}; }
 
   /// @brief Get the root path
   /// @details
   /// The root path is the root directory of the file system.
   /// @see get_mount_point() and get_partition_label()
   /// @return The root path
-  std::filesystem::path get_root_path() { return std::filesystem::path{get_mount_point()}; }
+  std::filesystem::path get_root_path() const { return std::filesystem::path{get_mount_point()}; }
+
+  /// @brief Convert file permissions to a string
+  /// @details This method converts file permissions to a string in the format "rwxrwxrwx".
+  /// @param permissions The file permissions
+  /// @return The file permissions as a string
+  std::string to_string(const std::filesystem::perms &permissions) const;
+
+  /// @brief Convert a time_t to a string
+  /// @details This method converts a time_t to a string in the format "Jan 01 00:00".
+  /// @param time The time_t to convert
+  /// @return The time as a string
+  std::string to_string(time_t time) const;
+
+  /// @brief Get the time of a file as a string
+  /// @details This method gets the time of a file as a string in the format "Jan 01 00:00".
+  /// @param path The path to the file
+  /// @return The time of the file as a string
+  /// @see file_time_to_string()
+  std::string get_file_time_as_string(const std::filesystem::path &path) const;
+
+  /// @brief Get a vector of files in a directory
+  /// @details This method returns a vector of paths to the files in a directory.
+  /// @param path The path to the directory
+  /// @param include_directories Whether to include directories in the output
+  /// @param recursive Whether to include files in subdirectories
+  /// @return A vector of paths to the files in the directory
+  std::vector<std::filesystem::path> get_files_in_path(const std::filesystem::path &path,
+                                                       bool include_directories = false,
+                                                       bool recursive = false);
 
   /// @brief List the contents of a directory
   /// @details
@@ -173,88 +203,7 @@ public:
   /// @param prefix The prefix to use for the output
   /// @return The contents of the directory
   std::string list_directory(const std::string &path, const ListConfig &config,
-                             const std::string &prefix = "") {
-    std::string result;
-    DIR *dir = opendir(path.c_str());
-    if (dir == nullptr) {
-      return result;
-    }
-    struct dirent *entry;
-    namespace fs = std::filesystem;
-    while ((entry = readdir(dir)) != nullptr) {
-      // skip the current and parent directories
-      if (strcmp(entry->d_name, ".") == 0 || strcmp(entry->d_name, "..") == 0) {
-        continue;
-      }
-      // use the config to determine output
-      std::error_code ec;
-      auto file_path = fs::path{path} / entry->d_name;
-      auto file_status = fs::status(file_path, ec);
-      if (ec) {
-        logger_.warn("Failed to get status for file: {}", file_path.string());
-      }
-      if (config.type) {
-        if (fs::is_directory(file_status)) {
-          result += "d";
-        } else if (fs::is_regular_file(file_status)) {
-          result += "-";
-        } else {
-          result += "?";
-        }
-      }
-      if (config.permissions) {
-        auto perms = file_status.permissions();
-        result += (perms & fs::perms::owner_read) != fs::perms::none ? "r" : "-";
-        result += (perms & fs::perms::owner_write) != fs::perms::none ? "w" : "-";
-        result += (perms & fs::perms::owner_exec) != fs::perms::none ? "x" : "-";
-        result += (perms & fs::perms::group_read) != fs::perms::none ? "r" : "-";
-        result += (perms & fs::perms::group_write) != fs::perms::none ? "w" : "-";
-        result += (perms & fs::perms::group_exec) != fs::perms::none ? "x" : "-";
-        result += (perms & fs::perms::others_read) != fs::perms::none ? "r" : "-";
-        result += (perms & fs::perms::others_write) != fs::perms::none ? "w" : "-";
-        result += (perms & fs::perms::others_exec) != fs::perms::none ? "x" : "-";
-        result += " ";
-      }
-      if (config.number_of_links) {
-        result += "1 ";
-      }
-      if (config.owner) {
-        result += "owner ";
-      }
-      if (config.group) {
-        result += "group ";
-      }
-      if (config.size) {
-        if (fs::is_regular_file(file_status)) {
-          result += fmt::format("{:>8} ", human_readable(fs::file_size(file_path, ec)));
-        } else {
-          result += fmt::format("{:>8} ", "");
-        }
-      }
-      if (config.date_time) {
-        auto ftime = fs::last_write_time(file_path, ec);
-        if (ec) {
-          result += "Jan 01 00:00 ";
-        } else {
-          // NOTE: std::chrono::system_clock::to_time_t is not implemented in ESP-IDF
-          // auto cftime = std::chrono::system_clock::to_time_t(ftime);
-          auto cftime = to_time_t(ftime);
-          std::tm tm = *std::localtime(&cftime);
-          char buffer[80];
-          std::strftime(buffer, sizeof(buffer), "%b %d %H:%M", &tm);
-          result += fmt::format("{:>12} ", buffer);
-        }
-      }
-      result += prefix;
-      result += entry->d_name;
-      result += "\r\n";
-      if (config.recursive && fs::is_directory(file_status)) {
-        result += list_directory(file_path, config, std::string{entry->d_name} + "/");
-      }
-    }
-    closedir(dir);
-    return result;
-  }
+                             const std::string &prefix = "");
 
   /// Function to convert a time_point to a time_t.
   /// \details This function converts a time_point to a time_t. This function
@@ -273,6 +222,9 @@ public:
   }
 
 protected:
+  static bool read_only_;     ///< Whether the file system is read only
+  static bool grow_on_mount_; ///< Whether to grow the file system on mount
+
   /// @brief Constructor
   /// @details
   /// The constructor is private to ensure that the class is a singleton.
@@ -285,44 +237,6 @@ protected:
   /// @details
   /// This method initializes the file system. It is protected and called only by the constructor.
   /// It is responsible for mounting the file system and creating the root directory.
-  void init() {
-    logger_.debug("Initializing file system");
-    esp_err_t err;
-
-    // Initialize LittleFS
-    esp_vfs_littlefs_conf_t conf = {
-        .base_path = get_mount_point().c_str(),
-        .partition_label = get_partition_label(),
-        .format_if_mount_failed = true,
-        .dont_mount = false,
-    };
-    err = esp_vfs_littlefs_register(&conf);
-    switch (err) {
-    case ESP_OK:
-      logger_.debug("LittleFS initialized");
-      break;
-    case ESP_ERR_NOT_FOUND:
-      logger_.error("Failed to find partition with label %s", get_partition_label());
-      break;
-    case ESP_FAIL:
-      logger_.error("Failed to mount or format filesystem");
-      break;
-    case ESP_ERR_INVALID_STATE:
-      logger_.error("Failed to initialize LittleFS, invalid state");
-      break;
-    default:
-      logger_.error("Failed to initialize LittleFS, unknown error");
-      break;
-    }
-    // NOTE: this will raise an exception if the LittleFS is not initialized
-    ESP_ERROR_CHECK(err);
-
-    // Create root directory
-    std::filesystem::path root_path = get_root_path();
-    if (!std::filesystem::exists(root_path)) {
-      logger_.debug("Creating root directory");
-      std::filesystem::create_directory(root_path);
-    }
-  }
+  void init();
 };
 } // namespace espp
